fix(ci): resolve ModuleNotFoundError and missing CLI flags
Two workflow failures introduced by PR #589:

1. `ia_parquet_uploader.py` imported from `scripts.pipeline.ia_s3`, which
   is not on sys.path when causaganha is installed as a package (only src/
   is the package root). This caused every `uv run causaganha` invocation to
   fail with ModuleNotFoundError, breaking the "Show backfill status" step
   in collect-zips.yml (which runs with `if: always()`).

   Fix: promote ia_s3.py to src/causaganha/pipeline/ia_s3.py (canonical
   location inside the installed package) and replace scripts/pipeline/ia_s3.py
   with a thin re-export wrapper for backwards compat with other scripts.

2. collect-today.yml passed `--skip-if-mostly-complete` to djen_backup but
   the flag was never defined in the CLI, causing the Collect Today job to fail
   with "No such option".

   Fix: add `--skip-if-mostly-complete` boolean option to djen_backup/__main__.py
   and the corresponding `skip_if_mostly_complete` field to SyncConfig.

Also adds `pydantic-settings>=2.0.0` as an explicit dependency since
causaganha/config.py depends on it.
