+++ pyproject.toml
@@ -23,6 +23,7 @@ dependencies = [
     "typer>=0.12.0",
     "pydantic>=2.10.0",
     "pydantic-ai>=0.0.14",
+    "pydantic-settings>=2.0.0",
     "ibis-framework[duckdb]>=9.0",
     "httpx>=0.27",
     "structlog>=24.1",
